+++ source/B2B.Transactions.Infrastructure/Configuration/CompositionRoot.cs
@@ -24,9 +24,9 @@ using B2B.Transactions.Configuration.DataAccess;
 using B2B.Transactions.IncomingMessages;
 using B2B.Transactions.Infrastructure.Common;
 using B2B.Transactions.Infrastructure.Configuration.Authentication;
+using B2B.Transactions.Infrastructure.Configuration.InternalCommands;
 using B2B.Transactions.Infrastructure.DataAccess;
 using B2B.Transactions.Infrastructure.DataAccess.Transaction;
-using B2B.Transactions.Infrastructure.InternalCommands;
 using B2B.Transactions.Infrastructure.Messages;
 using B2B.Transactions.Infrastructure.OutgoingMessages;
 using B2B.Transactions.Infrastructure.Serialization;
