+++ source/B2B.Transactions.Infrastructure/Configuration/InternalCommands/ICommand.cs
@@ -12,23 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Threading.Tasks;
-using NodaTime;
+using MediatR;
 
-namespace B2B.Transactions.Infrastructure.InternalCommands
+namespace B2B.Transactions.Infrastructure.Configuration.InternalCommands
 {
+    #pragma warning disable CA1040
     /// <summary>
-    /// Service for scheduling and enqueueing internal commands for later processing
+    /// CQRS command object
     /// </summary>
-    public interface ICommandScheduler
+    public interface ICommand : IRequest
     {
-        /// <summary>
-        /// Schedules or enqueues a command
-        /// </summary>
-        /// <param name="command"></param>
-        /// <param name="scheduleDate"></param>
-        /// <typeparam name="TCommand"><see cref="InternalCommand"/></typeparam>
-        Task EnqueueAsync<TCommand>(TCommand command, Instant? scheduleDate = null)
-            where TCommand : InternalCommand;
     }
+#pragma warning restore
 }
